refactor: ArticlePage query 분리
